move existing tests to correct path
